requirements.txt
# Core dependencies
pdfplumber
python-docx
Pillow
openai
tqdm
notion-client

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
python-dotenv
